Act040_Main.java     - Adição de campos para bundle. @#$ Act040_Main_Presenter.java Act040_Main_Presenter_Impl.java     - Metodos para criação de estrutura de SO expressa. @#$ Act91_Contract.kt     - Refatoração e metodo de save. @#$ Act091_BottomSheet.kt     - Refatoração e recuperação de dados da tela para save. @#$ Act091_BottomSheet_Item_Adapter.kt     - Correção na lista de serviços por pacote. @#$ Act091_Item_Adapter.kt     - Refatoração. @#$ Act091_Main.kt     - Adição de metodo de save. @#$ Act091_Presenter.kt     - Save de pacote adicionado. @#$ BottomEvent.kt BottomListEvent.kt     - Refatoração e alteração em modificador. @#$ SoPackExpressPacksLocal.kt     - Serialização para passagem via bundle. @#$ SoPackExpressPacksLocalDao.kt     - Correção de nome de tabela e encadeamento de tabelas. @#$ SoPackExpressPacksLocalSql001.kt     - Seleção unica de pacotes. @#$ SoPackExpressPacksLocalSql002.kt     - Seleção de pacotes por SO. @#$ SoPackExpressPacksLocalSql003.kt     - Geração de seq para registro. @#$ SoPackExpressServicesLocalDao.kt     - Correção de nome de tabela, de colunas da PK e encadeamento de tabelas. @#$ SoPackExpressServicesLocalSql001.kt     - Seleção unica de serviços. @#$ SoPackExpressServicesLocalSql002.kt     - Seleção serviços do pacote. @#$ SoPackExpressServicesLocalSql003.kt     - Geração de seq para registro. @#$ ToolBox_Inf.java     - Adequação do tipo. @#$

+++ app/src/main/java/com/namoadigital/prj001/util/ToolBox_Inf.java
@@ -4075,7 +4075,7 @@ public class ToolBox_Inf {
         return false;
     }
 
-    public static String getExpressSOFileName(int contract_code, int product_code, int category_price_code, int site_code, int operation_code) {
+    public static String getExpressSOFileName(int contract_code, long product_code, int category_price_code, long site_code, long operation_code) {
         return contract_code + "_" + category_price_code + "_" + product_code + "_" + site_code + "_" + operation_code + ".json";
     }
